Three modifications to the system bluetooth audio settings to enable (seemingly) greater stability and compatibility, based on personal testing and forum comments.
Changes are:
1) Always let the phone be master for BT connections; in testing, this reduces
Bluetooth "flush" calls, and prevents the HSP/HFP association from
disconnecting.  A commit in early 2010 was made to change the Bluez default
from "true" to "false". This undoes that change.  I believe the prior change
is not needed any more because of the updated Bluez stack.

2) Enables HSP and HFP both. No reason not to support this (it's an additonal
profile, not a supplement).  Memory usage does not change, and it enables
better compatibility and performance with BT devices.

3) Allows three connections on the HFP/HSP channel. Found this from scraping
various forum posts, as something necessary to be compatible with some
car head-units. No functional change for me, but found no indication it
had bad side-effects, so may as well include it while we're touching this file.

Change-Id: Idffcbce0bfda67f03fb4aa4183cf2c0cb8a672fa