Update patch set 1
Patch Set 1:

Thanks Danny. I tested with the old .29 for the Click, just to verify, and it was the same issue there. Any negative effects for other stuff, by using that alignment workaround in the device init.rc? I think that's the only option for now, in order to keep this feature, if it's not something that can be fixed in the AVRCP 1.3 code.

Patch-set: 1