Update patch set 1
Patch Set 1:

I figured it would be nice to point out that Danny's patch on the pastebin fixes an issue in where the last letter on AVRCP metadata would be cut (example: "Let it be" showed as "Let it b"). That problem is fixed after applying the 2nd pastebin patch on Motorola Atrix.

Patch-set: 1