Update patch set 2
Patch Set 2:

Controlling works fine on the JVC now as well. Only one weird issue that I'm not sure is because of this patch, or the JVC. It's that the media info redraws on the JVC screen every 5 seconds, and sometimes vanishes. I'll need to test without this patch and see if it's the same.

Patch-set: 2