Update patch set 3
Patch Set 3:

All good. Turned out the JVC has a motion sensor, that I wasn't aware of. That's what caused the media info to go away after 5 seconds, and then reappear once the hand got near it. When I disabled the sensor, the info stayed permanently on the screen.

Patch-set: 3