Update patch set 5
Patch Set 5:

This patch adds the ListPlayerApplicationSettingAttributes processing (mandatory) and fixes the metadata response.

I now also found out what's up with Giulio's car stereo: Apparently the firmware of that car stereo is broken and drops the BT connection if an AVRCP 1.3 capable device is connected. If the length of the EVENTS_SUPPORTED GetCapabilities PDU is wrong (and thus the whole PDU is broken), the car stereo stops any AVRCP 1.3 action and thus the crashing code path in the radio isn't triggered.

I'm not sure what to do with this, maybe a property to disable the AVRCP 1.3 support entirely? I haven't explored yet how easy/hard that would be.

Patch-set: 5