Update patch set 5
Patch Set 5:

Arne,

yes, I think it would be possible to watch out for the device name 'Ford Audio' (both Giulio's and my coworker's car stereo use that, so it should be safe to assume it's used consistently) and block AVRCP 1.3 communication based on that (reject every command sent with 'invalid command').

There's one caveat, though: We still would advertise AVRCP 1.3 support in the AVRCP profile descriptor (because that's determine on startup on the adapter, not the device level) although we would no longer properly support it. But as this is a device-specific quirk, I think it shouldn't matter.

I would prefer to implement this in a separate patch set, though, because making broken car stereos work is not directly related to making our implementation spec compliant.

Patch-set: 5