Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This looks like it covers the issues I was concerned with. Is anyone able to verify that this patch causes MAP support to work where it did not work previously? The behavior I was noticing is that no RFCOMM connection was ever established to the bluetoothmas service running on rfcomm channel 16 due to a malformed SDP record.

Patch-set: 1